+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.3.8 [unreleased]
+- chore: Wait on identify before returning connection [PR 47]
 - feat(repo): Allow custom repo store [PR 46]
 
+[PR 47]: https://github.com/dariusc93/rust-ipfs/pull/47
 [PR 46]: https://github.com/dariusc93/rust-ipfs/pull/46
 
 # 0.3.7